// web/src/api/api-manifest.ts
// @/api/api-manifest.ts
// This file is generated by protoc-gen-api-catalog. Do not edit manually.

export const APIs = {
  user: {
    v1: {
      CenterService: {
        EditCenterAvatar: 'user.v1.CenterService/EditCenterAvatar',
        EditCenterInfo: 'user.v1.CenterService/EditCenterInfo',
        EditCenterPassword: 'user.v1.CenterService/EditCenterPassword',
        GetCenterInfo: 'user.v1.CenterService/GetCenterInfo',
      },
      DeptService: {
        AddDept: 'user.v1.DeptService/AddDept',
        CheckDeptDelete: 'user.v1.DeptService/CheckDeptDelete',
        DeleteDeptCascade: 'user.v1.DeptService/DeleteDeptCascade',
        GetDept: 'user.v1.DeptService/GetDept',
        GetDeptByName: 'user.v1.DeptService/GetDeptByName',
        GetDeptChild: 'user.v1.DeptService/GetDeptChild',
        GetDeptTop: 'user.v1.DeptService/GetDeptTop',
        UpdateDept: 'user.v1.DeptService/UpdateDept',
        UpdatePriorityDept: 'user.v1.DeptService/UpdatePriorityDept',
      },
      LogService: {
        GetLogList: 'user.v1.LogService/GetLogList',
      },
      RoleService: {
        AddRole: 'user.v1.RoleService/AddRole',
        CheckDeleteRole: 'user.v1.RoleService/CheckDeleteRole',
        DeleteRole: 'user.v1.RoleService/DeleteRole',
        GetRole: 'user.v1.RoleService/GetRole',
        GetRoleAll: 'user.v1.RoleService/GetRoleAll',
        GetRoleList: 'user.v1.RoleService/GetRoleList',
        UpdateRole: 'user.v1.RoleService/UpdateRole',
      },
      UserService: {
        AddUser: 'user.v1.UserService/AddUser',
        DeleteUser: 'user.v1.UserService/DeleteUser',
        GetAPIs: 'user.v1.UserService/GetAPIs',
        GetCaptcha: 'user.v1.UserService/GetCaptcha',
        GetLoginCrypto: 'user.v1.UserService/GetLoginCrypto',
        GetMenus: 'user.v1.UserService/GetMenus',
        GetOnlineUserList: 'user.v1.UserService/GetOnlineUserList',
        GetUser: 'user.v1.UserService/GetUser',
        GetUserList: 'user.v1.UserService/GetUserList',
        HeartBeatUser: 'user.v1.UserService/HeartBeatUser',
        Login: 'user.v1.UserService/Login',
        Logout: 'user.v1.UserService/Logout',
        OfflineUser: 'user.v1.UserService/OfflineUser',
        ResetUserPassword: 'user.v1.UserService/ResetUserPassword',
        UGetDept: 'user.v1.UserService/UGetDept',
        UpdateUser: 'user.v1.UserService/UpdateUser',
      },
    },
  },
} as const

export type APIType = typeof APIs